requirements.txt
# PDL-POC Dependencies

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# People Data Labs SDK
peopledatalabs>=3.0.0

# Environment
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

